requirements.txt
imutils==0.5.4
tensorflow==2.8.0
keras==2.8.0
matplotlib==3.5.1
numpy==1.22.3
opencv-python==4.1.2.30
Pillow==9.1.0
scipy==1.5.4
wget==3.2
absl-py
astor==0.8.0
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
h5py==2.10.0
Keras-Applications
Keras-Preprocessing
kiwisolver==1.1.0
Markdown==3.1.1
opt-einsum==3.1.0
protobuf
pyparsing==2.4.2
python-dateutil==2.8.0
PyYAML==5.1.2
six==1.12.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2

